% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{add_no_choice_to_design}
\alias{add_no_choice_to_design}
\title{Add no-choice option to design matrix}
\usage{
add_no_choice_to_design(design_matrix, n_alts)
}
\arguments{
\item{design_matrix}{Matrix of profileIDs (questions x alternatives)}

\item{n_alts}{Current number of alternatives per question}
}
\value{
Updated design matrix with no-choice alternatives added
}
\description{
Adds a no-choice alternative (profileID = 0) as the last alternative in each question
}
