% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{build_interaction_terms}
\alias{build_interaction_terms}
\title{Build interaction terms in logitr format}
\usage{
build_interaction_terms(interactions)
}
\description{
Build interaction terms in logitr format
}
