% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{build_parameter_vector_with_interactions}
\alias{build_parameter_vector_with_interactions}
\title{Build parameter vector including interaction values}
\usage{
build_parameter_vector_with_interactions(
  codedParNames,
  attrs,
  means,
  interactions
)
}
\description{
Build parameter vector including interaction values
}
