% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_inspect.R
\name{cbc_balance}
\alias{cbc_balance}
\title{Counts of attribute balance}
\usage{
cbc_balance(design, atts = NULL)
}
\arguments{
\item{design}{A data frame of a survey design.}

\item{atts}{A vector of attributes. Defaults to \code{NULL}, in which case all
of the column names of the \code{design} data frame are used except for those
that end in \code{"ID"}.}
}
\value{
A list showing counts of the number of choice questions that contain
the unique number of levels for each attribute.
}
\description{
This function prints out a summary of the counts of each level for each
attribute across all choice questions as well as the two-way counts across
all pairs of attributes for a given design.
}
\examples{
# Insert examples
}
