% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derror.R
\name{cbc_db_error}
\alias{cbc_db_error}
\alias{cbc_db_error.cbc_survey}
\alias{cbc_db_error.cbc_design}
\alias{cbc_db_error.data.frame}
\title{Compute DB-error for a choice experiment design with random parameters}
\usage{
cbc_db_error(x, ...)

\method{cbc_db_error}{cbc_survey}(x, priors = NULL, exclude = NULL, ...)

\method{cbc_db_error}{cbc_design}(x, priors = NULL, exclude = NULL, ...)

\method{cbc_db_error}{data.frame}(x, priors = NULL, exclude = NULL, ...)
}
\arguments{
\item{x}{Either a \code{cbc_design} object created by \code{cbc_design()} or a data frame
containing a choice experiment design}

\item{...}{Additional arguments passed to methods}

\item{priors}{A \code{cbc_priors} object created by \code{cbc_priors()} that includes
random parameters. If \code{x} is a \code{cbc_design} object, the priors from that
object will be used by default.}

\item{exclude}{Character vector of attribute names to exclude from DB-error
calculation}
}
\value{
The DB-error value
}
\description{
Compute DB-error for a choice experiment design with random parameters
}
\details{
DB-error (Bayesian D-error) accounts for uncertainty in parameter estimates
by averaging D-error across multiple parameter draws. This is only meaningful
when priors include random parameters with parameter draws.
}
\examples{
# Create profiles with random priors
profiles <- cbc_profiles(
  price = c(1, 2, 3),
  type = c("A", "B", "C")
)

priors <- cbc_priors(
  profiles = profiles,
  price = rand_spec(dist = "n", mean = -0.5, sd = 0.2),
  type = c(0.2, 0.3)
)

design <- cbc_design(
  profiles = profiles,
  priors = priors,
  n_alts = 2,
  n_q = 4,
  method = "sequential"
)

# Compute DB-error
cbc_db_error(design)
}
