% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{cbc_inspect_balance}
\alias{cbc_inspect_balance}
\alias{cbc_inspect_balance.cbc_design}
\alias{cbc_inspect_balance.data.frame}
\alias{cbc_balance}
\title{Inspect attribute balance in design}
\usage{
cbc_inspect_balance(x, ...)

\method{cbc_inspect_balance}{cbc_design}(x, ...)

\method{cbc_inspect_balance}{data.frame}(x, ...)

cbc_balance(x, ...)
}
\arguments{
\item{x}{Either a \code{cbc_design} object created by \code{cbc_design()} or a data frame
containing a choice experiment design}

\item{...}{Additional arguments passed to methods}
}
\value{
Invisibly returns the balance statistics, but primarily prints
formatted information to the console
}
\description{
This function prints out a summary of the individual and pairwise counts of
each level for each attribute across all choice questions in the design.
}
\examples{
library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make a survey design from all possible profiles
design <- cbc_design(
  profiles = profiles,
  n_alts   = 3,   # Number of alternatives per question
  n_q      = 6    # Number of questions per respondent
)

# Inspect the design balance
cbc_inspect_balance(design)
}
\keyword{balance}
\keyword{logit}
\keyword{logitr}
\keyword{mixed}
\keyword{mnl}
\keyword{mxl}
\keyword{overlap}
