% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{cbc_inspect_overlap}
\alias{cbc_inspect_overlap}
\alias{cbc_inspect_overlap.cbc_design}
\alias{cbc_inspect_overlap.data.frame}
\alias{cbc_overlap}
\title{Inspect attribute overlap in design}
\usage{
cbc_inspect_overlap(x, ...)

\method{cbc_inspect_overlap}{cbc_design}(x, ...)

\method{cbc_inspect_overlap}{data.frame}(x, ...)

cbc_overlap(x, ...)
}
\arguments{
\item{x}{Either a \code{cbc_design} object created by \code{cbc_design()} or a data frame
containing a choice experiment design}

\item{...}{Additional arguments passed to methods}
}
\value{
Invisibly returns the overlap statistics, but primarily prints
formatted information to the console
}
\description{
This function prints out a summary of the amount of "overlap" across
attributes within the choice questions. For example, for each attribute, the
count under \code{"1"} is the number of choice questions in which the same level
was shown across all alternatives for that attribute (because there was only
one level shown). Likewise, the count under \code{"2"} is the number of choice
questions in which only two unique levels of that attribute were shown, and
so on.
}
\examples{
library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  freshness = c("Excellent", "Average", "Poor"),
  type      = c("Fuji", "Gala", "Honeycrisp")
)

# Make a randomized survey design
design <- cbc_design(
  profiles = profiles,
  n_alts   = 3, # Number of alternatives per question
  n_q      = 6 # Number of questions per respondent
)

# Inspect the design overlap
cbc_inspect_overlap(design)
}
