% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_inspect.R
\name{cbc_overlap}
\alias{cbc_overlap}
\title{Counts of attribute overlap}
\usage{
cbc_overlap(design, atts = NULL)
}
\arguments{
\item{design}{A data frame of a survey design.}

\item{atts}{A vector of attributes. Defaults to \code{NULL}, in which case all
of the column names of the \code{design} data frame are used except for those
that end in \code{"ID"}.}
}
\value{
A list showing counts of the number of choice questions that contain
the unique number of levels for each attribute.
}
\description{
This function helps understand the degree of overlap for each level of each
attribute in a given survey design. The list that is returned shows the
number of choice questions that contain a given number of unique levels. For
example, the number under \code{"1"} means the number of questions where the
same attribute level was shown across all alternatives (because there was
only one level shown).
}
\examples{
# Insert examples
}
