% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_power.R
\name{cbc_power}
\alias{cbc_power}
\title{Estimate the same model on different size subsets of data}
\usage{
cbc_power(
  nbreaks = 10,
  n_q = 1,
  data,
  outcome,
  obsID,
  pars,
  rand_pars = NULL,
  return_models = FALSE,
  panelID = NULL,
  clusterID = NULL,
  robust = FALSE,
  predict = FALSE,
  n_cores = NULL,
  ...
)
}
\arguments{
\item{nbreaks}{The number of different sample size groups.}

\item{n_q}{Number of questions per respondent. Defaults to \code{1} if not
specified.}

\item{data}{The data, formatted as a \code{data.frame} object.}

\item{outcome}{The name of the column that identifies the outcome variable,
which should be coded with a \code{1} for \code{TRUE} and \code{0} for \code{FALSE}.}

\item{obsID}{The name of the column that identifies each observation.}

\item{pars}{The names of the parameters to be estimated in the model.
Must be the same as the column names in the \code{data} argument.}

\item{rand_pars}{A named vector whose names are the random parameters and
values the distribution: \code{'n'} for normal or \code{'ln'} for log-normal.
Defaults to \code{NULL}.}

\item{return_models}{If \code{TRUE}, a list of all estimated models is returned.
This can be useful if you want to extract other outputs from each model,
such as the variance-covariance matrix, etc. Defaults to \code{FALSE}.}

\item{panelID}{The name of the column that identifies the individual (for
panel data where multiple observations are recorded for each individual).
Defaults to \code{NULL}.}

\item{clusterID}{The name of the column that identifies the cluster
groups to be used in model estimation. Defaults to \code{NULL}.}

\item{robust}{Determines whether or not a robust covariance matrix is
estimated. Defaults to \code{FALSE}. Specification of a \code{clusterID} will override
the user setting and set this to `TRUE' (a warning will be displayed in this
case). Replicates the functionality of Stata's cmcmmixlogit.}

\item{predict}{If \code{TRUE}, predicted probabilities, fitted values, and
residuals are also included in the returned model objects. Defaults to
\code{FALSE}.}

\item{n_cores}{The number of cores to use for parallel processing.
Set to \code{1} to run serially Defaults to \code{NULL}, in which case the number of
cores is set to \code{parallel::detectCores() - 1}. Max cores allowed is capped
at \code{parallel::detectCores()}.}

\item{...}{Other arguments that are passed to \code{logitr::logitr()} for model
estimation. See the \code{logitr} documentation for details about other
available arguments.}
}
\description{
This function estimates the same model multiple times using different size
subsets of a set of choice data. The number of models to run is set by the
\code{nbreaks} argument, which breaks up the data into groups of increasing
sample sizes. All models are estimated models using the {logitr} package.
}
\examples{
# Insert examples
}
\keyword{logit,}
\keyword{logitr,}
\keyword{mnl,}
\keyword{mxl,}
\keyword{power}
\keyword{sample}
\keyword{size,}
