% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_design.R
\name{cbc_profiles}
\alias{cbc_profiles}
\title{Make a data frame of all combinations of attribute levels}
\usage{
cbc_profiles(levels)
}
\arguments{
\item{levels}{A named list of vectors defining each attribute (the names)
and each level for each attribute (the vectors). For example, a design
with two attributes, \code{"price"} (numeric) and \code{"type"} (categorical),
that each had three levels could be defined as
\code{levels = list(price = c(1, 2, 3), type = c("type1", "type2", "type3"))}.}
}
\value{
A data frame of all possible combinations of attribute levels.
}
\description{
This function creates a data frame of of all possible combinations of
attribute levels defined in the \code{levels} argument.
}
\examples{
library(cbcTools)

# A simple conjoint experiment about apples

# Define the attributes and levels
levels <- list(
  price     = seq(1, 4, 0.5), # $ per pound
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c("Excellent", "Average", "Poor")
)

# Generate all profiles with all attributes dummy-coded
profiles <- cbc_profiles(levels)
}
