% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{cbc_restrict}
\alias{cbc_restrict}
\title{Obtain a restricted set of profiles}
\usage{
cbc_restrict(profiles, ...)
}
\arguments{
\item{profiles}{A data frame in which each row is a possible profile.
This can be generated using the \code{cbc_profiles()} function.}

\item{...}{Any number of lists defining pairs of restricted attribute levels.
Each restriction should be provided as a named list defining one restricted pair,
where the item name is the attribute and the value is the level, e.g.
\code{list(type = 'Fuji', freshness = 'Poor')}. Separate each list by a comma.}
}
\value{
A restricted set of profiles as a data frame.
}
\description{
This function returns a restricted set of profiles as a data frame.
}
\examples{
library(cbcTools)

# Generate all profiles for a simple conjoint experiment about apples
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Obtain a restricted subset of profiles
profiles_restricted <- cbc_restrict(
    profiles,
    list(type = 'Fuji', freshness = 'Poor'),
    list(price = 3, type = 'Gala')
)
}
