% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{compute_both_d_errors_from_survey}
\alias{compute_both_d_errors_from_survey}
\title{Compute both D-errors from a full survey design data frame}
\usage{
compute_both_d_errors_from_survey(survey_design, opt_env)
}
\arguments{
\item{survey_design}{Full survey design data frame}

\item{opt_env}{Optimization environment}
}
\value{
List with null_d_error and prior_d_error (if available)
}
\description{
Compute both D-errors from a full survey design data frame
}
