% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design2.R
\name{compute_d_error_from_probs}
\alias{compute_d_error_from_probs}
\title{Compute D-error from X matrix and probabilities}
\usage{
compute_d_error_from_probs(X_design, probs, obsID)
}
\arguments{
\item{X_design}{Design matrix (all alternatives x parameters)}

\item{probs}{Vector of choice probabilities}

\item{obsID}{Vector indicating which observation each row belongs to}
}
\value{
D-error value
}
\description{
Compute D-error from X matrix and probabilities
}
