% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derror.R
\name{compute_db_error}
\alias{compute_db_error}
\title{Compute DB-error using Monte Carlo simulation}
\usage{
compute_db_error(X_list, mu, sigma, n_draws)
}
\description{
Compute DB-error using Monte Carlo simulation
}
