% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{compute_design_d_error}
\alias{compute_design_d_error}
\title{Compute D-error for entire design using profileID matrix}
\usage{
compute_design_d_error(design_matrix, opt_env)
}
\arguments{
\item{design_matrix}{Matrix of profileIDs (questions x alternatives)}

\item{opt_env}{Optimization environment}
}
\value{
Numeric D-error value
}
\description{
Reconstructs X matrix from profileIDs and computes full information matrix
}
