% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{compute_design_probabilities_draw}
\alias{compute_design_probabilities_draw}
\title{Compute probabilities for entire design (single draw)}
\usage{
compute_design_probabilities_draw(design_matrix, opt_env, draw_index)
}
\arguments{
\item{design_matrix}{Matrix of profileIDs}

\item{opt_env}{Optimization environment}

\item{draw_index}{Which parameter draw to use}
}
\value{
Vector of probabilities for all alternatives
}
\description{
Compute probabilities for entire design (single draw)
}
