% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{compute_design_probabilities_fixed}
\alias{compute_design_probabilities_fixed}
\title{Compute probabilities for entire design (fixed parameters)}
\usage{
compute_design_probabilities_fixed(design_matrix, opt_env)
}
\arguments{
\item{design_matrix}{Matrix of profileIDs}

\item{opt_env}{Optimization environment}
}
\value{
Vector of probabilities for all alternatives
}
\description{
Compute probabilities for entire design (fixed parameters)
}
