% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design2.R
\name{compute_info_matrix}
\alias{compute_info_matrix}
\title{Compute information matrix from X and probability lists}
\usage{
compute_info_matrix(X_list, P_list)
}
\arguments{
\item{X_list}{List of X matrices (one per choice question)}

\item{P_list}{List of probability vectors (one per choice question)}
}
\value{
List with M_list and M_total
}
\description{
Compute information matrix from X and probability lists
}
