% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{compute_question_probabilities}
\alias{compute_question_probabilities}
\title{Compute choice probabilities for a question}
\usage{
compute_question_probabilities(question_profiles, opt_env)
}
\arguments{
\item{question_profiles}{Vector of profileIDs}

\item{opt_env}{Optimization environment}
}
\value{
Vector of choice probabilities
}
\description{
Compute choice probabilities for a question
}
