% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{decode_categorical_variables}
\alias{decode_categorical_variables}
\title{Decode categorical variables from dummy coding}
\usage{
decode_categorical_variables(design, categorical_structure)
}
\arguments{
\item{design}{Design data frame with dummy-coded variables}

\item{categorical_structure}{Information about original categorical structure}
}
\value{
Design data frame with categorical variables restored
}
\description{
Decode categorical variables from dummy coding
}
