% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDoe.R
\name{evaluateDoe}
\alias{evaluateDoe}
\title{Evaluate a design of experiment}
\usage{
evaluateDoe(doe)
}
\arguments{
\item{doe}{A design of experiment data frame. Each row is an alternative,
each column is an attribute.}
}
\value{
Returns a list of information about a given design, including the
D-efficiency and whether or not the design is balanced.
\code{levels} argument.
}
\description{
Obtain a list of information about a given design, including the
D-efficiency and whether or not the design is balanced.
}
\examples{
# Define the attributes and levels
levels <- list(
  price     = seq(1, 4, 0.5), # $ per pound
  type      = c('Fuji', 'Gala', 'Honeycrisp', 'Pink Lady', 'Red Delicious'),
  freshness = c('Excellent', 'Average', 'Poor')
)

# Make a fraction-factorial design of experiment based on D-efficiency
doe <- makeDoe(levels, type = "D", nTrials = 70)

# Print the design efficiency and whether it is balanced:
evaluateDoe(doe)
}
