% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{extract_interaction_value}
\alias{extract_interaction_value}
\title{Extract interaction value for a specific coded parameter}
\usage{
extract_interaction_value(param_name, interactions, attrs)
}
\description{
Extract interaction value for a specific coded parameter
}
