% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{finalize_design_object}
\alias{finalize_design_object}
\title{Finalize the design object with metadata including both D-errors}
\usage{
finalize_design_object(
  design,
  opt_env,
  design_result,
  method,
  n_alts,
  n_q,
  n_resp,
  n_blocks,
  no_choice,
  label,
  randomize_questions,
  randomize_alts
)
}
\arguments{
\item{design}{The design data frame (always dummy-coded)}

\item{opt_env}{Optimization environment}

\item{design_result}{Result from design generation}

\item{method}{Design method used}

\item{n_alts}{Number of alternatives}

\item{n_q}{Number of questions per respondent}

\item{n_resp}{Number of respondents}

\item{n_blocks}{Number of blocks}

\item{no_choice}{No-choice option included}

\item{label}{Label variable used}

\item{randomize_questions}{Question randomization setting}

\item{randomize_alts}{Alternative randomization setting}
}
\value{
cbc_design object
}
\description{
Finalize the design object with metadata including both D-errors
}
