% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{find_problematic_questions}
\alias{find_problematic_questions}
\title{Find all questions that have problems}
\usage{
find_problematic_questions(design_matrix, opt_env)
}
\description{
Find all questions that have problems
}
