% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{generate_initial_random_matrix}
\alias{generate_initial_random_matrix}
\title{Generate initial random design matrix}
\usage{
generate_initial_random_matrix(opt_env)
}
\description{
Generate initial random design matrix
}
