% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{generate_random_design}
\alias{generate_random_design}
\title{Generate a random design using profileID sampling with matrix operations}
\usage{
generate_random_design(opt_env)
}
\description{
Generate a random design using profileID sampling with matrix operations
}
