% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design2.R
\name{generate_sequential_design}
\alias{generate_sequential_design}
\title{Placeholder for sequential design generation}
\usage{
generate_sequential_design(
  opt_env,
  n_alts,
  n_q,
  n_blocks,
  max_iter,
  n_start,
  max_attempts
)
}
\description{
Placeholder for sequential design generation
}
