% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{get_categorical_structure_from_profiles}
\alias{get_categorical_structure_from_profiles}
\title{Get categorical structure information from original profiles}
\usage{
get_categorical_structure_from_profiles(profiles, attr_names)
}
\arguments{
\item{profiles}{Original profiles data frame}

\item{attr_names}{Attribute names}
}
\value{
List containing information about categorical variables
}
\description{
Get categorical structure information from original profiles
}
