% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_expected_interaction_params}
\alias{get_expected_interaction_params}
\title{Get all possible interaction parameter names that logitr might create}
\usage{
get_expected_interaction_params(interactions, attrs)
}
\description{
Get all possible interaction parameter names that logitr might create
}
