% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_var_names_with_interactions}
\alias{get_var_names_with_interactions}
\title{Get variable names including interaction terms for logitr}
\usage{
get_var_names_with_interactions(profile_attrs, interactions = NULL)
}
\description{
Get variable names including interaction terms for logitr
}
