% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{has_dominant_alternative}
\alias{has_dominant_alternative}
\title{Check if a question has a dominant alternative}
\usage{
has_dominant_alternative(question_profiles, opt_env)
}
\arguments{
\item{question_profiles}{Vector of profileIDs}

\item{opt_env}{Optimization environment}
}
\value{
Logical indicating if dominance exists
}
\description{
Check if a question has a dominant alternative
}
