% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design2.R
\name{has_duplicate_question}
\alias{has_duplicate_question}
\title{Check if a question is a duplicate of previous questions}
\usage{
has_duplicate_question(question_profiles, design_matrix, question_index)
}
\arguments{
\item{question_profiles}{Vector of profileIDs for current question}

\item{design_matrix}{Current design matrix}

\item{question_index}{Current question index}
}
\value{
Logical indicating if question is duplicate
}
\description{
Check if a question is a duplicate of previous questions
}
