% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{has_partial_dominance}
\alias{has_partial_dominance}
\title{Check for partial dominance between alternatives}
\usage{
has_partial_dominance(question_profiles, opt_env)
}
\arguments{
\item{question_profiles}{Vector of profileIDs in this question}

\item{opt_env}{Optimization environment containing partial_utilities matrix}
}
\value{
Logical indicating if partial dominance exists
}
\description{
Uses pre-computed partial utilities matrix for fast dominance checking
Checks if any alternative has higher partial utility for ALL attributes
}
