% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_design.R
\name{infer_categorical_structure}
\alias{infer_categorical_structure}
\title{Infer categorical structure from data}
\usage{
infer_categorical_structure(data, encoding)
}
\arguments{
\item{data}{A data.frame}

\item{encoding}{Current encoding type}
}
\value{
A list of categorical structure information
}
\description{
Infer categorical structure from data
}
\keyword{internal}
