% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_design.R
\name{infer_from_coded_columns}
\alias{infer_from_coded_columns}
\title{Infer categorical structure from dummy/effects coded columns}
\usage{
infer_from_coded_columns(data, attr_cols, encoding)
}
\arguments{
\item{data}{A data.frame}

\item{attr_cols}{Column names to analyze}

\item{encoding}{Encoding type}
}
\value{
A list of categorical structure
}
\description{
Infer categorical structure from dummy/effects coded columns
}
\keyword{internal}
