% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.cbc_choices}
\alias{is.cbc_choices}
\title{Check if object is a cbc_choices object}
\usage{
is.cbc_choices(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a cbc_choices object
}
\description{
Check if object is a cbc_choices object
}
