% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.cbc_design}
\alias{is.cbc_design}
\title{Check if object is a cbc_design object}
\usage{
is.cbc_design(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a cbc_design object
}
\description{
Check if object is a cbc_design object
}
