% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{is.cbc_priors}
\alias{is.cbc_priors}
\title{Check if object is a cbc_priors object}
\usage{
is.cbc_priors(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a cbc_priors object
}
\description{
Check if object is a cbc_priors object
}
