% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{is.cbc_profiles}
\alias{is.cbc_profiles}
\title{Check if object is a cbc_profiles object}
\usage{
is.cbc_profiles(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a cbc_profiles object
}
\description{
Check if object is a cbc_profiles object
}
