% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.cbc_survey}
\alias{is.cbc_survey}
\title{Check if object is a cbc_survey object}
\usage{
is.cbc_survey(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if x is a cbc_survey object
}
\description{
Check if object is a cbc_survey object
}
