% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{is_dummy_coded}
\alias{is_dummy_coded}
\title{Check if design is dummy-coded}
\usage{
is_dummy_coded(design)
}
\arguments{
\item{design}{Design object to check}
}
\value{
Logical indicating if design is dummy-coded
}
\description{
Check if design is dummy-coded
}
