% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{is_interaction_parameter}
\alias{is_interaction_parameter}
\title{Check if a parameter name represents an interaction}
\usage{
is_interaction_parameter(param_name)
}
\description{
Check if a parameter name represents an interaction
}
