% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design2.R
\name{is_valid_question}
\alias{is_valid_question}
\title{Check if a question (set of profileIDs) is valid}
\usage{
is_valid_question(question_profiles, opt_env, design_matrix, question_index)
}
\arguments{
\item{question_profiles}{Vector of profileIDs for this question}

\item{opt_env}{Optimization environment}

\item{design_matrix}{Current design matrix}

\item{question_index}{Current question index}
}
\value{
Logical indicating if question is valid
}
\description{
Check if a question (set of profileIDs) is valid
}
