% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSurvey.R
\name{makeSurvey}
\alias{makeSurvey}
\title{Make a randomized conjoint survey from a design of experiment}
\usage{
makeSurvey(
  doe,
  nResp,
  nAltsPerQ,
  nQPerResp,
  outsideGood = FALSE,
  label = NULL,
  group
)
}
\arguments{
\item{doe}{A design of experiment data frame. Each row is an alternative,
each column is an attribute.}

\item{nResp}{Maximum number of survey respondents.}

\item{nAltsPerQ}{Number of alternatives per question.}

\item{nQPerResp}{Number of questions per respondent.}

\item{outsideGood}{Include an outside good in the choice sets? Defaults to
\code{FALSE}. If included, the total number of alternatives per question will be
one more than the provided \code{nAltsPerQ} argument.}

\item{label}{The name of the variable to use in a "labeled" design
such that each set of alternatives contains one of each of the levels in
the \code{label} attribute. If included, the \code{nAltsPerQ} argument will be
ignored as its value is determined by the unique number of levels in the
\code{label} variable. Defaults to \code{NULL}.}

\item{group}{No longer used as of v0.0.9 - if provided, this is passed
to the \code{label} argument and a warning is displayed.}
}
\value{
Returns a conjoint survey by randomly sampling from a design of
experiment data frame. The sampling is done iteratively to remove the
possibility of duplicate alternatives appearing in the same choice question.
}
\description{
Make a randomized conjoint survey from a design of experiment
}
\examples{
# Define the attributes and levels
levels <- list(
  price     = seq(1, 4, 0.5), # $ per pound
  type      = c('Fuji', 'Gala', 'Honeycrisp', 'Pink Lady', 'Red Delicious'),
  freshness = c('Excellent', 'Average', 'Poor')
)

# Make a full-factorial design of experiment
doe <- makeDoe(levels)

# Re-code levels
doe <- recodeDoe(doe, levels)

# Make the conjoint survey by randomly sampling from the doe
survey <- makeSurvey(
  doe       = doe,  # Design of experiment
  nResp     = 2000, # Total number of respondents (upper bound)
  nAltsPerQ = 3,    # Number of alternatives per question
  nQPerResp = 6     # Number of questions per respondent
)
}
