% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{matches_interaction_spec}
\alias{matches_interaction_spec}
\title{Check if a coded parameter matches an interaction specification}
\usage{
matches_interaction_spec(param_name, param_parts, int_spec, attrs)
}
\description{
Check if a coded parameter matches an interaction specification
}
