% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{miscmethods.cbc_errors}
\alias{miscmethods.cbc_errors}
\alias{plot.cbc_errors}
\title{Methods for cbc_errors objects}
\usage{
\method{plot}{cbc_errors}(x, ...)
}
\arguments{
\item{x}{is an object of class \code{cbc_errors}.}

\item{...}{further arguments.}
}
\description{
Miscellaneous methods for \code{cbc_errors} class objects.
}
