% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{miscmethods.cbc_models}
\alias{miscmethods.cbc_models}
\alias{print.cbc_models}
\title{Methods for cbc_models objects}
\usage{
\method{print}{cbc_models}(
  x,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  ...
)
}
\arguments{
\item{x}{is an object of class \code{cbc_models}.}

\item{digits}{the number of digits for printing, defaults to \code{3}.}

\item{width}{the width of the printing.}

\item{...}{further arguments.}
}
\value{
No return value, prints a summary of estimated models.
}
\description{
Miscellaneous methods for \code{cbc_models} class objects.
}
