% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{miscmethods.cbc_power_errors}
\alias{miscmethods.cbc_power_errors}
\alias{plot.cbc_power_errors}
\title{Methods for cbc_power_errors objects}
\usage{
\method{plot}{cbc_power_errors}(x, ...)
}
\arguments{
\item{x}{is an object of class \code{cbc_power_errors}.}

\item{...}{further arguments.}
}
\value{
Returns a ggplot2 object plotting standard errors versus sample
size.
}
\description{
Miscellaneous methods for \code{cbc_power_errors} class objects.
}
