% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{miscmethods.cjmodels}
\alias{miscmethods.cjmodels}
\alias{print.cjmodels}
\title{Methods for cjmodels objects}
\usage{
\method{print}{cjmodels}(
  x,
  digits = max(3, getOption("digits") - 2),
  width = getOption("width"),
  ...
)
}
\arguments{
\item{x}{is an object of class \code{cjmodels}.}

\item{digits}{the number of digits for printing, defaults to \code{3}.}

\item{width}{the width of the printing.}

\item{...}{further arguments.}
}
\description{
Miscellaneous methods for \code{cjmodels} class objects.
}
