% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.cbc_power}
\alias{plot.cbc_power}
\title{Plot method for cbc_power objects}
\usage{
\method{plot}{cbc_power}(x, type = "power", power_threshold = 0.8, ...)
}
\arguments{
\item{x}{A cbc_power object}

\item{type}{Type of plot: "power" for power curves or "se" for standard error curves}

\item{power_threshold}{Power threshold for horizontal reference line (only for power plots). Defaults to 0.8}

\item{...}{Additional arguments passed to ggplot}
}
\description{
Plot method for cbc_power objects
}
