% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{plot_compare_power}
\alias{plot_compare_power}
\title{Plot a comparison of different design powers}
\usage{
plot_compare_power(...)
}
\arguments{
\item{...}{Any number of data frame containing power results obtained from
the \code{cbc_power()} function, separated by commas.}
}
\value{
A plot comparing the power curves of different designs.
}
\description{
This function creates a ggplot2 object comparing the power curves of
different designs. Each design is color coded and each facet (sub plot)
is a model coefficient.
}
