% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cbc_choices}
\alias{print.cbc_choices}
\title{Print method for cbc_choices objects}
\usage{
\method{print}{cbc_choices}(x, ...)
}
\arguments{
\item{x}{A cbc_choices object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for cbc_choices objects
}
