% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{print.cbc_comparison}
\alias{print.cbc_comparison}
\alias{print.cbc_error_comparison}
\title{Print method for comparisons}
\usage{
\method{print}{cbc_comparison}(x, ...)

\method{print}{cbc_error_comparison}(x, ...)
}
\arguments{
\item{x}{A cbc_comparison or cbc_error_comparison object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for comparisons
}
