% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derror.R, R/methods.R
\name{print.cbc_d_error_comparison}
\alias{print.cbc_d_error_comparison}
\title{Print method for D-error comparisons}
\usage{
\method{print}{cbc_d_error_comparison}(x, ...)

\method{print}{cbc_d_error_comparison}(x, ...)
}
\arguments{
\item{x}{A cbc_d_error_comparison object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for D-error comparisons

Print method for D-error comparisons
}
