% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{print.cbc_design_comparison}
\alias{print.cbc_design_comparison}
\title{Print method for design comparisons}
\usage{
\method{print}{cbc_design_comparison}(x, ...)
}
\arguments{
\item{x}{A cbc_design_comparison object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for design comparisons
}
