% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{print.cbc_design_flagged}
\alias{print.cbc_design_flagged}
\title{Print method for flagged designs}
\usage{
\method{print}{cbc_design_flagged}(x, ...)
}
\arguments{
\item{x}{A cbc_design_flagged object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for flagged designs
}
