% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{print.cbc_errors}
\alias{print.cbc_errors}
\title{Enhanced print method for cbc_errors objects}
\usage{
\method{print}{cbc_errors}(x, ...)
}
\arguments{
\item{x}{A cbc_errors object}

\item{...}{Additional arguments passed to print}
}
\description{
Enhanced print method for cbc_errors objects
}
