% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{print.cbc_inspection}
\alias{print.cbc_inspection}
\title{Print method for cbc_inspection objects}
\usage{
\method{print}{cbc_inspection}(x, ...)
}
\arguments{
\item{x}{A cbc_inspection object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for cbc_inspection objects
}
