% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cbc_power}
\alias{print.cbc_power}
\title{Print method for cbc_power objects}
\usage{
\method{print}{cbc_power}(x, ...)
}
\arguments{
\item{x}{A cbc_power object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for cbc_power objects
}
