% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cbc_power_errors}
\alias{print.cbc_power_errors}
\title{Enhanced print method for cbc_power_errors objects}
\usage{
\method{print}{cbc_power_errors}(x, ...)
}
\arguments{
\item{x}{A cbc_power_errors object}

\item{...}{Additional arguments passed to print}
}
\description{
Enhanced print method for cbc_power_errors objects
}
