% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{print.cbc_profiles}
\alias{print.cbc_profiles}
\title{Print method for cbc_profiles objects}
\usage{
\method{print}{cbc_profiles}(x, ...)
}
\arguments{
\item{x}{A cbc_profiles object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for cbc_profiles objects
}
