% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cbc_survey}
\alias{print.cbc_survey}
\title{Enhanced print method for cbc_survey objects}
\usage{
\method{print}{cbc_survey}(x, ...)
}
\arguments{
\item{x}{A cbc_survey object}

\item{...}{Additional arguments passed to print}
}
\description{
Enhanced print method for cbc_survey objects
}
