% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{rand_spec}
\alias{rand_spec}
\title{Create a random parameter specification}
\usage{
rand_spec(dist = "n", mean, sd, correlations = NULL)
}
\arguments{
\item{dist}{Character. Distribution type: "n" for normal, "ln" for log-normal,
or "cn" for censored normal}

\item{mean}{Numeric. Mean parameter value(s)}

\item{sd}{Numeric. Standard deviation parameter value(s)}

\item{correlations}{List of correlation specifications created by cor_spec()}
}
\value{
A random parameter specification list
}
\description{
Create a random parameter specification
}
\examples{
# For a continuous parameter
rand_spec(dist = "ln", mean = -0.5, sd = 0.4)

# For a categorical parameter with correlations
rand_spec(
  dist = "n",
  mean = c("Fuji" = 0.2, "Gala" = 0.3),
  sd = c("Fuji" = 0.4, "Gala" = 0.4),
  correlations = list(
    cor_spec(with = "price", value = 0.3),
    cor_spec(with = "type", level = "Fuji", with_level = "Gala", value = 0.5)
  )
)
}
