% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_choices.R
\name{randN}
\alias{randN}
\title{Define a prior (assumed) model parameter as normally-distributed.}
\usage{
randN(mu = 0, sigma = 1)
}
\arguments{
\item{mu}{Vector of means, defaults to \code{0}.}

\item{sigma}{Vector of standard deviations, defaults to \code{1}.}
}
\value{
A list defining normally-distributed parameters of the prior
(assumed) utility model used to simulate choices in the \code{cbc_choices()}
function.
}
\description{
Define a prior (assumed) model parameter as normally-distributed.
Used in the \code{cbc_choices()} function.
}
\examples{
# Insert example
}
