% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design2.R
\name{randomize_question_order}
\alias{randomize_question_order}
\title{Randomize question order within a respondent's design}
\usage{
randomize_question_order(resp_design, n_q, n_alts)
}
\arguments{
\item{resp_design}{Design for one respondent}

\item{n_q}{Number of questions per respondent}

\item{n_alts}{Number of alternatives per question}
}
\value{
Design with randomized question order
}
\description{
Randomize question order within a respondent's design
}
