% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_design.R
\name{reconstruct_design_attributes}
\alias{reconstruct_design_attributes}
\title{Reconstruct minimal design attributes from a data frame}
\usage{
reconstruct_design_attributes(data)
}
\arguments{
\item{data}{A data.frame with valid CBC design structure}
}
\value{
The data with reconstructed attributes
}
\description{
Reconstruct minimal design attributes from a data frame
}
\keyword{internal}
