% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{replace_single_choice_set}
\alias{replace_single_choice_set}
\title{Replace a single choice set with new random draws}
\usage{
replace_single_choice_set(design, qID, profiles, n_alts, label = NULL)
}
\arguments{
\item{design}{The design data frame}

\item{qID}{The question ID to replace}

\item{profiles}{The profiles to sample from}

\item{n_alts}{Number of alternatives per question}

\item{label}{Label variable for labeled designs}
}
\value{
Updated design with the specified choice set replaced
}
\description{
Replace a single choice set with new random draws
}
\keyword{internal}
