% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{sample_labeled_profiles}
\alias{sample_labeled_profiles}
\title{Sample profiles for labeled design}
\usage{
sample_labeled_profiles(label_constraints, n_alts)
}
\arguments{
\item{label_constraints}{Label constraint structure}

\item{n_alts}{Number of alternatives (should match number of label groups)}
}
\value{
Vector of profileIDs, one from each label group
}
\description{
Sample profiles for labeled design
}
