% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{sample_question_profiles}
\alias{sample_question_profiles}
\title{Sample profileIDs for a single question}
\usage{
sample_question_profiles(opt_env, n_alts)
}
\arguments{
\item{opt_env}{Optimization environment}

\item{n_alts}{Number of alternatives to sample}
}
\value{
Vector of profileIDs
}
\description{
Sample profileIDs for a single question
}
