% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design2.R
\name{sample_valid_question}
\alias{sample_valid_question}
\title{Sample a valid question while avoiding conflicts}
\usage{
sample_valid_question(
  opt_env,
  n_alts,
  design_matrix,
  question_index,
  n_q,
  n_resp
)
}
\arguments{
\item{opt_env}{Optimization environment}

\item{n_alts}{Number of alternatives}

\item{design_matrix}{Current design matrix}

\item{question_index}{Which question to replace}

\item{n_q}{Number of questions per respondent}

\item{n_resp}{Number of respondents}
}
\value{
Vector of profileIDs for the question
}
\description{
Sample a valid question while avoiding conflicts
}
