% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{set_interaction_values}
\alias{set_interaction_values}
\title{Set interaction parameter values in the parameter vector}
\usage{
set_interaction_values(pars_mean, codedParNames, interactions, attrs)
}
\description{
Set interaction parameter values in the parameter vector
}
