% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{setup_label_constraints}
\alias{setup_label_constraints}
\title{Set up label constraints for labeled designs}
\usage{
setup_label_constraints(profiles, label, n_alts)
}
\description{
Set up label constraints for labeled designs
}
