% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cbc_design}
\alias{summary.cbc_design}
\title{Detailed summary method for cbc_design objects}
\usage{
\method{summary}{cbc_design}(object, ...)
}
\arguments{
\item{object}{A cbc_design object}

\item{...}{Additional arguments passed to summary}
}
\description{
Detailed summary method for cbc_design objects
}
