% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{summary.cbc_power}
\alias{summary.cbc_power}
\title{Summary method for cbc_power objects}
\usage{
\method{summary}{cbc_power}(object, power_threshold = 0.8, ...)
}
\arguments{
\item{object}{A cbc_power object}

\item{power_threshold}{Minimum power threshold to report sample size requirements}

\item{...}{Additional arguments}
}
\description{
Summary method for cbc_power objects
}
