% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{validate_interaction_levels}
\alias{validate_interaction_levels}
\title{Validate level specifications for interactions}
\usage{
validate_interaction_levels(int, attrs)
}
\description{
Validate level specifications for interactions
}
