% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_design.R
\name{validate_or_reconstruct}
\alias{validate_or_reconstruct}
\title{User-friendly wrapper to validate design with error messages}
\usage{
validate_or_reconstruct(
  data,
  require_class = FALSE,
  allow_choices = TRUE,
  context = "this function"
)
}
\arguments{
\item{data}{Input data}

\item{require_class}{Whether to require proper class}

\item{allow_choices}{Whether to allow cbc_choices objects}

\item{context}{Character describing where this is being called from (for error messages)}
}
\value{
Validated/reconstructed data, or stops with error
}
\description{
User-friendly wrapper to validate design with error messages
}
\keyword{internal}
