% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{validate_priors}
\alias{validate_priors}
\title{Validate that priors are compatible with profiles}
\usage{
validate_priors(priors, profiles, no_choice)
}
\description{
Validate that priors are compatible with profiles
}
