% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{validate_priors_profiles}
\alias{validate_priors_profiles}
\title{Validate that priors are compatible with profiles}
\usage{
validate_priors_profiles(priors, profiles)
}
\arguments{
\item{priors}{A cbc_priors object}

\item{profiles}{A cbc_profiles object}
}
\value{
Invisibly returns TRUE if compatible, throws error or warning if not
}
\description{
This function checks if priors were created for the same profiles structure
}
