% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_inspect.R
\name{cbc_balance}
\alias{cbc_balance}
\title{Counts of attribute balance}
\usage{
cbc_balance(design, atts = NULL)
}
\arguments{
\item{design}{A data frame of a survey design.}

\item{atts}{A vector of attributes. Defaults to \code{NULL}, in which case all
of the column names of the \code{design} data frame are used except for those
that end in \code{"ID"}.}
}
\value{
A list showing counts of the number of choice questions that contain
the unique number of levels for each attribute.
}
\description{
This function prints out a summary of the counts of each level for each
attribute across all choice questions as well as the two-way counts across
all pairs of attributes for a given design.
}
\examples{
library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make a randomized survey design
design <- cbc_design(
  profiles = profiles,
  n_resp   = 300, # Number of respondents
  n_alts   = 3, # Number of alternatives per question
  n_q      = 6 # Number of questions per respondent
)

# Inspect the design balance
cbc_balance(design)

# Inspect the design overlap
cbc_overlap(design)
}
