% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_choices.R
\name{cbc_choices}
\alias{cbc_choices}
\title{Simulate choices for a survey design}
\usage{
cbc_choices(design, obsID = "obsID", priors = NULL, n_draws = 100)
}
\arguments{
\item{design}{A data frame of a survey design.}

\item{obsID}{The name of the column in \code{design} that identifies each choice
observation. Defaults to \code{"obsID"}.}

\item{priors}{A list of one or more prior parameters that define a prior
(assumed) utility model used to simulate choices for the \code{survey} data frame.
If \code{NULL} (the default), choices will be randomly assigned.}

\item{n_draws}{The number of Halton draws to use for simulated choices
for mixed logit models. Defaults to \code{100}.}
}
\value{
Returns the \code{design} data frame with an additional \code{choice} column
identifying the simulated choices.
}
\description{
Simulate choices for a survey design, either randomly or according to a
utility model defined by user-provided prior parameters.
}
\examples{
library(cbcTools)

# A simple conjoint experiment about apples

# Define the attributes and levels
levels <- list(
  price     = seq(1, 4, 0.5), # $ per pound
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c("Excellent", "Average", "Poor")
)

# Generate all all possible profiles
profiles <- cbc_profiles(levels)

# Make a randomized survey design
design <- cbc_design(
  profiles = profiles,
  n_resp   = 300, # Number of respondents
  n_alts   = 3, # Number of alternatives per question
  n_q      = 6 # Number of questions per respondent
)
}
