% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices.R
\name{cbc_choices}
\alias{cbc_choices}
\alias{cbc_choices.cbc_survey}
\alias{cbc_choices.data.frame}
\title{Simulate choices for a survey design}
\usage{
cbc_choices(x, ...)

\method{cbc_choices}{cbc_survey}(x, priors = NULL, n_draws = 100, ...)

\method{cbc_choices}{data.frame}(x, priors = NULL, n_draws = 100, ...)
}
\arguments{
\item{x}{Either a \code{cbc_survey} object created by \code{cbc_survey()} or a data frame
containing a survey design}

\item{...}{Additional arguments passed to methods}

\item{priors}{A \code{cbc_priors} object created by \code{cbc_priors()}, or \code{NULL} for
random choices. If \code{x} is a \code{cbc_survey} object, the priors from that
object will be used by default.}

\item{n_draws}{The number of Halton draws to use for simulated choices
for mixed logit models. Defaults to \code{100}.}
}
\value{
Returns the input data with an additional \code{choice} column
identifying the simulated choices.
}
\description{
Simulate choices for a survey design, either randomly or according to a
utility model defined by user-provided prior parameters. All choices are
simulated using the 'logitr' package. For more details see the JSS article
on the 'logitr' package (Helveston, 2023).
}
\examples{
library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make a survey design from all possible profiles
design <- cbc_design(
  profiles = profiles,
  n_alts   = 3,   # Number of alternatives per question
  n_q      = 6    # Number of questions per respondent
)

# Create survey and simulate choices
survey <- cbc_survey(design, n_resp = 100)
data_with_choices <- cbc_choices(survey)

# Simulate choices according to a prior utility model
priors <- cbc_priors(
  profiles = profiles,
  price     = -0.1,
  type      = c(0.1, 0.2),
  freshness = c(0.1, 0.2)
)

data <- cbc_choices(survey, priors = priors)

# Simulate choices according to a prior utility model with random parameters
priors_random <- cbc_priors(
  profiles = profiles,
  price = -0.1,
  type = rand_spec(mean = c(0.1, 0.2), sd = c(0.5, 0.8)),
  freshness = c(0.1, 0.2)
)

data <- cbc_choices(survey, priors = priors_random)
}
\references{
Helveston, J. P. (2023). logitr: Fast Estimation of Multinomial and Mixed Logit Models with Preference Space and Willingness-to-Pay Space Utility Parameterizations. Journal of Statistical Software, 105(10), 1–37,
\doi{10.18637/jss.v105.i10}
}
\keyword{logit}
\keyword{logitr}
\keyword{mixed}
\keyword{mnl}
\keyword{mxl}
\keyword{simulation}
