% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{cbc_decode}
\alias{cbc_decode}
\title{Convert dummy-coded CBC data back to categorical format}
\usage{
cbc_decode(data)
}
\arguments{
\item{data}{A \code{cbc_design} or \code{cbc_choices} object with dummy-coded categorical variables}
}
\value{
The input object with categorical variables restored to their original format
}
\description{
This function is depreciated. Use \code{cbc_encode()} instead
}
