% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{cbc_design}
\alias{cbc_design}
\title{Make a choice-based conjoint survey design}
\usage{
cbc_design(
  profiles,
  n_resp,
  n_alts,
  n_q,
  n_blocks = 1,
  n_draws = 100,
  n_start = 5,
  no_choice = FALSE,
  label = NULL,
  method = "random",
  randomize_questions = TRUE,
  randomize_alts = TRUE,
  priors = NULL,
  prior_no_choice = NULL,
  max_iter = 50,
  parallel = FALSE
)
}
\arguments{
\item{profiles}{A data frame in which each row is a possible profile. This
can be generated using the \code{cbc_profiles()} function.}

\item{n_resp}{Number of survey respondents.}

\item{n_alts}{Number of alternatives per choice question.}

\item{n_q}{Number of questions per respondent.}

\item{n_blocks}{Number of blocks used D-efficient or Bayesian D-efficient
designs. Max allowable is one block per respondent. Defaults to \code{1},
meaning every respondent sees the same choice set.}

\item{n_draws}{Number of draws used in simulating the prior distribution used
in Bayesian D-efficient designs. Defaults to \code{100}.}

\item{n_start}{A numeric value indicating the number of random start designs
to use in obtaining a D-efficient or Bayesian D-efficient design.
The default is \code{1}. Increasing \code{n_start} can result in a more efficient
design at the expense of increased computational time.}

\item{no_choice}{Include a "no choice" option in the choice sets? Defaults to
\code{FALSE}. If \code{TRUE}, the total number of alternatives per question will be
one more than the provided \code{n_alts} argument.}

\item{label}{The name of the variable to use in a "labeled" design (also
called an "alternative-specific" design) such that each set of alternatives
contains one of each of the levels in the \code{label} attribute. If used, the
\code{n_alts} argument will be ignored as its value is defined by the unique
number of levels in the \code{label} attribute Defaults to \code{NULL}.}

\item{method}{Choose the design method to use. Options include \code{"random"}
and \code{"sequential"}. Defaults to \code{"random"}. See details below for complete
description of each method.}

\item{priors}{A list of one or more assumed prior parameters used to generate
a D-efficient or Bayesian D-efficient design. Defaults to \code{NULL}}

\item{prior_no_choice}{Prior utility value for the "no choice" alternative.
Only required if \code{no_choice = TRUE}. Defaults to \code{NULL}.}

\item{max_iter}{A numeric value indicating the maximum number allowed
iterations when searching for a D-efficient or Bayesian D-efficient design.
The default is 50.}

\item{randomize}{Logical. If \code{TRUE}, both the choice question order and the
order of the alternatives within a given choice question will be
randomized across each respondent. Does not apply to design created using
the \code{"random"} method. Defaults to \code{TRUE}.}
}
\value{
The returned \code{design} data frame contains a choice-based conjoint
survey design where each row is an alternative. It includes the following
columns:
\itemize{
\item \code{profileID}: Identifies the profile in \code{profiles}.
\item \code{blockID}: If blocking is used, identifies each unique block.
\item \code{respID}: Identifies each survey respondent.
\item \code{qID}: Identifies the choice question answered by the respondent.
\item \code{altID}: Identifies the alternative in any one choice observation.
\item \code{obsID}: Identifies each unique choice observation across all respondents.
}
}
\description{
This function creates a data frame containing a choice-based conjoint survey
design where each row is an alternative. Generate a variety of survey
designs, including random design, D-efficient designs, and Bayesian
D-efficient designs as well as designs with "no choice" options, blocking,
and "labeled" designs (also known as "alternative specific" designs).
}
\details{
The \code{method} argument determines the design method used. Options
are:
\itemize{
\item \code{"random"}: Creates a design by randomly sampling from profiles
\item \code{"sequential"}: Creates a D-efficient design using sequential improvement. Bayesian D-efficient designs can be obtained by specifying a prior model with the \code{priors} argument that include a covariance matrix.

All methods ensure that the two following criteria are met:
\enumerate{
\item No two profiles are the same within any one choice set.
\item No two choice sets are the same within any one respondent.
}

The table below summarizes method compatibility with other design options:\tabular{lllll}{
   Method \tab No choice? \tab Labeled designs? \tab Restricted profiles? \tab Blocking? \cr
   \code{"random"} \tab Yes \tab Yes \tab Yes \tab No \cr
   \code{"sequential"} \tab No \tab Yes \tab Yes \tab Yes \cr
}

}

The \code{"sequential"} method creates a design by sequentially improving D-efficiency:
\enumerate{
\item Start with a random design
\item Compute initial D-error
\item For each question, alternative, and attribute:
\itemize{
\item Try all possible level changes to the attribute.
\item Keep changes that improve the D-error.
}
\item Repeat until no further improvements or max iterations are reached
}
}
\examples{
library(cbcTools)

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make a random survey design (default)
design_random <- cbc_design(
  profiles = profiles,
  n_resp   = 100, # Number of respondents
  n_alts   = 3,   # Number of alternatives per question
  n_q      = 6    # Number of questions per respondent
)

# Create priors
priors <- cbc_priors(
  price = -0.5,
  type = c(0.2, 0.3),
  freshness = c(0.4, 0.8)
)

# Make a D-efficient design with priors using the "sequential" method
design <- cbc_design(
  profiles = profiles,
  n_resp = 100,
  n_alts = 3,
  n_q = 6,
  method = "sequential",
  priors = priors
)
}
\keyword{DoE.base}
\keyword{design}
\keyword{experiment}
\keyword{idefix}
\keyword{logit}
\keyword{logitr}
\keyword{mixed}
\keyword{mnl}
\keyword{mxl}
