% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{cbc_design}
\alias{cbc_design}
\title{Make a random or D-efficient choice-based conjoint survey design}
\usage{
cbc_design(
  profiles,
  n_resp,
  n_alts,
  n_q,
  no_choice = FALSE,
  label = NULL,
  d_eff = FALSE
)
}
\arguments{
\item{profiles}{A data frame in which each row is a possible profile.
This can be generated using the \code{cbc_profiles()} function.}

\item{n_resp}{Number of survey respondents.}

\item{n_alts}{Number of alternatives per choice question.}

\item{n_q}{Number of questions per respondent.}

\item{no_choice}{Include a "none" option in the choice sets? Defaults to
\code{FALSE}. If \code{TRUE}, the total number of alternatives per question will be
one more than the provided \code{n_alts} argument.}

\item{label}{The name of the variable to use in a "labeled" design
(also called an "alternative-specific design") such that each set of
alternatives contains one of each of the levels in the \code{label} attribute.
If used, the \code{n_alts} argument will be ignored as its value is defined by
the unique number of levels in the \code{label} variable. Defaults to \code{NULL}.}

\item{d_eff}{If \code{TRUE}, returns a D-efficient design where each respondent
is shown the same design, otherwise a fully randomized design is returned.
Defaults to \code{FALSE}. \strong{This feature is not yet implemented}}
}
\value{
A data frame containing a choice-based conjoint survey design where
each row is an alternative.
}
\description{
This function creates a data frame containing a choice-based conjoint survey
design where each row is an alternative. Designs can be either fully
randomized or D-efficient, in which case an implementation of the
Modified Federov algorithm is used via the \code{idefix} package (still in
development).
}
\examples{
library(cbcTools)

# A simple conjoint experiment about apples

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make a randomized survey design
design_rand <- cbc_design(
  profiles = profiles,
  n_resp   = 300, # Number of respondents
  n_alts   = 3, # Number of alternatives per question
  n_q      = 6 # Number of questions per respondent
)

# Make a randomized survey design with a "no choice" option
design_rand_nochoice <- cbc_design(
  profiles  = profiles,
  n_resp    = 300, # Number of respondents
  n_alts    = 3, # Number of alternatives per question
  n_q       = 6, # Number of questions per respondent
  no_choice = TRUE
)

# Make randomized labeled survey design with each "type" appearing in each
# choice question
design_rand_labeled <- cbc_design(
  profiles  = profiles,
  n_resp    = 300, # Number of respondents
  n_alts    = 3, # Number of alternatives per question
  n_q       = 6, # Number of questions per respondent
  label     = "type"
)
}
