% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_inspect.R
\name{cbc_overlap}
\alias{cbc_overlap}
\title{Counts of attribute overlap}
\usage{
cbc_overlap(design, atts = NULL)
}
\arguments{
\item{design}{A data frame of a survey design.}

\item{atts}{A vector of attributes. Defaults to \code{NULL}, in which case all
of the column names of the \code{design} data frame are used except for those
that end in \code{"ID"}.}
}
\value{
A list showing counts of the number of choice questions that contain
the unique number of levels for each attribute.
}
\description{
This function prints out a summary of the amount of "overlap" across
attributes within the choice questions. For example, for each attribute, the
count under \code{"1"} is the number of choice questions in which the same level
was shown across all alternatives for that attribute (because there was only
one level shown). Likewise, the count under \code{"2"} is the number of choice
questions in which only two unique levels of that attribute were shown, and
so on.
}
\examples{
# Insert examples
}
