% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_design.R
\name{cbc_profiles}
\alias{cbc_profiles}
\title{Make a coded matrix of all possible combinations of attribute levels}
\usage{
cbc_profiles(levels, coding = NULL)
}
\arguments{
\item{levels}{A named list of vectors defining each attribute (the names)
and each level for each attribute (the vectors). For example, a design
with two attributes, \code{"price"} and \code{"type"}, that each had three levels
should be defined as
\code{levels = list(price = c(1, 2, 3), type = c("1", "2", "3"))}.}

\item{coding}{The type of coding to be used for each attribute. Valid
arguments for are \code{"C"} for continuous attributes, \code{"D"} for dummy-coded
attributes, and \code{"E"} for effects-coded attributes When using \code{"C"} the
values provided for the associated attribute will be directly used in the
design matrix. Defaults to \code{NULL}, in which case all attributes are
dummy-coded.}
}
\value{
A coded matrix of all possible combinations of attribute levels.
}
\description{
This function is a wrapper for the \code{Profiles()} function from the \code{idefix}
package, which creates a coded matrix of all possible combinations of
attribute levels defined in the \code{levels} argument.
}
\examples{
# Define the attributes and levels
levels <- list(
  price     = seq(1, 4, 0.5), # $ per pound
  type      = c('Fuji', 'Gala', 'Honeycrisp', 'Pink Lady', 'Red Delicious'),
  freshness = c('Excellent', 'Average', 'Poor')
)

# Generate all profiles with all attributes dummy-coded
profiles <- cbc_profiles(levels)

# Generate all profiles with price as a continuous variable
profiles <- cbc_profiles(levels, coding = c("C", "D", "D"))
}
