% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbc_design.R
\name{cbc_profiles}
\alias{cbc_profiles}
\title{Make a data frame of all combinations of attribute levels}
\usage{
cbc_profiles(levels)
}
\arguments{
\item{levels}{A named list of vectors defining each attribute (the names)
and each level for each attribute (the vectors). Conditional levels can
also be included by setting each level of an attribute to another named list
that determines the levels for other attributes for that specific level.}
}
\value{
A data frame of all possible combinations of attribute levels.
}
\description{
This function creates a data frame of of all possible combinations of
attribute levels defined in the \code{levels} argument.
}
\examples{
library(cbcTools)

# A simple conjoint experiment about apples

# Define the attributes and levels
levels <- list(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  freshness = c("Excellent", "Average", "Poor"),
  type      = c("Fuji", "Gala", "Honeycrisp")
)

# Generate profiles
profiles <- cbc_profiles(levels)

# Define attributes and levels containing conditional levels
levels <- list(
  price = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  freshness = c("Excellent", "Average", "Poor"),
  type = list(
    "Fuji" = list(
        price = c(2, 2.5, 3)
    ),
    "Gala" = list(
        price = c(1, 1.5, 2)
    ),
    "Honeycrisp" = list(
        price = c(2.5, 3, 3.5, 4, 4.5, 5),
        freshness = c("Excellent", "Average")
    )
  )
)

# Generate profiles
profiles <- cbc_profiles(levels)

}
