% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{cbc_profiles}
\alias{cbc_profiles}
\title{Make a data frame of all combinations of attribute levels}
\usage{
cbc_profiles(...)
}
\arguments{
\item{...}{Any number of named vectors defining each attribute and their levels,
e.g. \code{price = c(1, 2, 3)}. Separate each vector by a comma.}
}
\value{
A data frame of all possible combinations of attribute levels with
class \code{cbc_profiles}.
}
\description{
This function creates a data frame of of all possible combinations of
attribute levels.
}
\examples{
library(cbcTools)

# Generate all profiles for a simple conjoint experiment about apples
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)
}
