% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{cbc_survey}
\alias{cbc_survey}
\title{Create a survey from a conjoint design}
\usage{
cbc_survey(design, n_resp, randomize_questions = TRUE, randomize_alts = TRUE)
}
\arguments{
\item{design}{A cbc_design object created by \code{cbc_design()}. If the
\code{"random"} method was used to create the design, then the design will
be ignored and a fully randomized survey will be generated where each
respondent sees a unique, random design.}

\item{n_resp}{Number of survey respondents}

\item{randomize_questions}{Logical. Whether to randomize question order for each respondent}

\item{randomize_alts}{Logical. Whether to randomize alternative order within questions}
}
\value{
A data frame with class cbc_survey containing the complete survey design
}
\description{
This function takes a conjoint design and replicates it across a specified
number of respondents to create a complete survey design.
}
