% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{cor_spec}
\alias{cor_spec}
\title{Create a correlation specification for random parameters}
\usage{
cor_spec(with, value, level = NULL, with_level = NULL)
}
\arguments{
\item{with}{Character. Name of attribute to correlate with}

\item{value}{Numeric. Correlation value between -1 and 1}

\item{level}{Character. For categorical variables, specific level to correlate from}

\item{with_level}{Character. For categorical variables, specific level to correlate with}
}
\value{
A correlation specification list
}
\description{
Create a correlation specification for random parameters
}
