% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{plot_compare_power}
\alias{plot_compare_power}
\title{Plot a comparison of different design powers}
\usage{
plot_compare_power(...)
}
\arguments{
\item{...}{Any number of data frame containing power results obtained from
the \code{cbc_power()} function, separated by commas.}
}
\value{
A plot comparing the power curves of different designs.
}
\description{
This function creates a ggplot2 object comparing the power curves of
different designs. Each design is color coded and each facet (sub plot)
is a model coefficient.
}
\examples{
\dontrun{
library(cbcTools)

# Generate all possible profiles
profiles <- cbc_profiles(
  price     = c(1, 1.5, 2, 2.5, 3),
  type      = c("Fuji", "Gala", "Honeycrisp"),
  freshness = c('Poor', 'Average', 'Excellent')
)

# Make designs to compare: random vs sequential
design_random <- cbc_design(
  profiles = profiles,
  n_alts = 3, n_q = 6, method = "random"
)

# Same priors will be used in sequential design and simulated choices
priors <- cbc_priors(
  profiles = profiles,
  price     = -0.1,
  type      = c(0.1, 0.2),
  freshness = c(0.1, 0.2)
)

design_sequential <- cbc_design(
  profiles  = profiles,
  priors = priors,
  n_alts = 3, n_q = 6, method = "sequential"
)

# Create surveys and simulate choices
survey_random <- cbc_survey(design_random, n_resp = 100)
choices_random <- cbc_choices(survey_random, priors = priors)

survey_sequential <- cbc_survey(design_sequential, n_resp = 100)
choices_sequential <- cbc_choices(survey_sequential, priors = priors)

# Obtain power for each design
power_random <- cbc_power(choices_random, nbreaks = 5, n_cores = 2)
power_sequential <- cbc_power(choices_sequential, nbreaks = 5, n_cores = 2)

# Compare power of each design
plot_compare_power(power_random, power_sequential)
}
}
