% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cbc_comparison}
\alias{print.cbc_comparison}
\title{Print method for cbc_comparison objects}
\usage{
\method{print}{cbc_comparison}(x, ...)
}
\arguments{
\item{x}{A cbc_comparison object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for cbc_comparison objects
}
