% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cbc_design}
\alias{print.cbc_design}
\title{Concise print method for cbc_design objects}
\usage{
\method{print}{cbc_design}(x, ...)
}
\arguments{
\item{x}{A cbc_design object}

\item{...}{Additional arguments passed to print}
}
\description{
Concise print method for cbc_design objects
}
