% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.cbc_errors}
\alias{print.cbc_errors}
\title{Print method for design error metrics}
\usage{
\method{print}{cbc_errors}(x, ...)
}
\arguments{
\item{x}{A cbc_errors object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for design error metrics
}
