% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design2.R
\name{randomize_alternative_order}
\alias{randomize_alternative_order}
\title{Randomize alternative order within questions}
\usage{
randomize_alternative_order(resp_design, n_q, n_alts)
}
\arguments{
\item{resp_design}{Design for one respondent}

\item{n_q}{Number of questions per respondent}

\item{n_alts}{Number of alternatives per question}
}
\value{
Design with randomized alternative order
}
\description{
Randomize alternative order within questions
}
